Added stb_image.h for loading images

+++ include/r_slide.h
@@ -30,9 +30,9 @@ typedef struct rSlide {
 
 rSlide rslide_create(const char* filepath);
 void rslide_delete(rSlide* slide);
-rText rtext_create(const char* text, int x, int y, int font_size, unsigned int color);
+rText rtext_create(const char* text, float x, float y, int font_size, unsigned int color);
 void rtext_delete(rText* text);
-rImage rimage_create(const char* filepath, int x, int y, int width, int height);
+rImage rimage_create(const char* filepath, float x, float y, int width, int height);
 void rimage_delete(rImage* image);
 
 #endif // R_SLIDE_H
